Ejercicios Bases de datos 2: muevo las comprobacionesa bloques separados de consulta principal de cada página

+++ ejercicios/bases-datos/bases-de-datos-2/bases-de-datos-2-1/borrar-2.php
@@ -13,10 +13,18 @@ cabecera("Borrar 2", MENU_VOLVER);
 
 $id = recoge("id", []);
 
+$idSeleccionadosOk = false;
+
 if (count($id) == 0) {
     print "    <p class=\"aviso\">No se ha seleccionado ningún registro.</p>\n";
 } else {
+    $idSeleccionadosOk = true;
+}
+
+if ($idSeleccionadosOk) {
     foreach ($id as $indice => $valor) {
+        $registroEncontradoOk = false;
+
         $consulta = "SELECT COUNT(*) FROM $cfg[tablaPersonas]
                      WHERE id = :indice";
 
@@ -28,6 +36,10 @@ if (count($id) == 0) {
         } elseif ($resultado->fetchColumn() == 0) {
             print "    <p class=\"aviso\">Registro no encontrado.</p>\n";
         } else {
+            $registroEncontradoOk = true;
+        }
+
+        if ($registroEncontradoOk) {
             $consulta = "DELETE FROM $cfg[tablaPersonas]
                          WHERE id = :indice";
